// schemas.d.ts
export interface GQLUser {
  id: string;
  email: string;
  name: string;
  username: string;
  bio?: string;
  visitedCountries?: Array<string | null>;
  phoneNumber?: GQLPhoneNumber;
  created_at?: string;
  image?: string;
  images?: Array<string | null>;
  banned?: boolean;
  location?: GQLLocation;
  friends?: Array<GQLUser | null>;
  mutualFriendsCount?: number;
  friendshipStatus?: GQLFriendshipType;
  notificationRange?: number;
  expoToken?: string;
  tags?: Array<string | null>;
  interests?: Array<GQLActivity | null>;
  home?: GQLHome;
  languages?: Array<string | null>;
  gender?: GQLGender;
  birthDate?: GQLBirthDate;
  website?: string;
}

export interface GQLPhoneNumber {
  countryCode?: string;
  digits?: string;
}

export interface GQLLocation {
  city?: string;
  lat?: number;
  lng?: number;
  country?: string;
  displayName?: string;
  placeId?: string;
  fullAddress?: string;
}

export const enum GQLFriendshipType {
  FRIEND = 'FRIEND',
  REQUESTED = 'REQUESTED',
  REQUESTED_BY = 'REQUESTED_BY',
  BLOCKED = 'BLOCKED',
  BLOCKED_BY = 'BLOCKED_BY',
  STRANGER = 'STRANGER'
}

export const enum GQLActivity {
  HIKING = 'HIKING',
  SCUBADIVING = 'SCUBADIVING',
  SURFING = 'SURFING',
  SKIING = 'SKIING',
  SNOWBOARDING = 'SNOWBOARDING',
  DANCING = 'DANCING',
  SAILING = 'SAILING',
  YOGA = 'YOGA',
  BEACH = 'BEACH',
  RESTAURANTS = 'RESTAURANTS',
  COOKING_CLASSES = 'COOKING_CLASSES',
  LOCAL_MARKETS = 'LOCAL_MARKETS',
  MUSEUMS = 'MUSEUMS',
  HISTORY_TOURS = 'HISTORY_TOURS',
  CULTURAL_FESTIVAL = 'CULTURAL_FESTIVAL',
  WILDLIFE_TOURS = 'WILDLIFE_TOURS',
  EXPLORING_NATURE = 'EXPLORING_NATURE',
  WINE_TASTING = 'WINE_TASTING',
  PARTYING = 'PARTYING',
  CONCERTS = 'CONCERTS',
  LANGUAGE_CLASSES = 'LANGUAGE_CLASSES',
  VOLUNTEERING = 'VOLUNTEERING'
}

export interface GQLHome {
  value?: GQLLocation;
  visibilityLevel?: GQLVisibilityLevel;
}

export const enum GQLVisibilityLevel {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
  FRIENDS_ONLY = 'FRIENDS_ONLY'
}

export interface GQLGender {
  value?: string;
  visibilityLevel?: GQLVisibilityLevel;
}

export interface GQLBirthDate {
  value?: string;
  visibilityLevel?: GQLVisibilityLevel;
}
// ...
export interface GQLPageInput {
  count: number;
  page?: number;
  cursor?: string;
}

export interface GQLSort {
  field?: GQLField;
  direction?: GQLDirection;
}

export const enum GQLField {
  UPDATED_AT = 'UPDATED_AT',
  CREATED_AT = 'CREATED_AT',
  START_DATE = 'START_DATE'
}

export const enum GQLDirection {
  ASC = 'ASC',
  DESC = 'DESC'
}

export interface GQLGetPlansResponse {
  results?: Array<GQLPlan | null>;
  page?: GQLPage;
}

export interface GQLPage {
  pageSize?: number;
  hasNext?: boolean;
  total?: number;
  nextCursor?: string;
}
// ...
export interface GQLMessage {
  id?: string;
  channelId?: string;
  senderId?: string;
  data?: string;
  messageType?: GQLMessageType;
  createdAt?: string;
  updatedAt?: string;
  repliesCount?: number;
  replyToMessageId?: string;
  attachment?: GQLAttachment;
}

export const enum GQLMessageType {
  TEXT = 'TEXT',
  MEDIA = 'MEDIA',
  GIF = 'GIF',
  PLAN_REFERENCE = 'PLAN_REFERENCE',
  SYSTEM_MESSAGE = 'SYSTEM_MESSAGE'
}

export type GQLAttachment = GQLMediaAttachment | GQLPlanAttachment;

/** Use this to resolve union type Attachment */
export type GQLPossibleAttachmentTypeNames =
'MediaAttachment' |
'PlanAttachment';

export interface GQLAttachmentNameMap {
  Attachment: GQLAttachment;
  MediaAttachment: GQLMediaAttachment;
  PlanAttachment: GQLPlanAttachment;
}

export interface GQLMediaAttachment {
  data?: Array<GQLMediaItem | null>;
}

export interface GQLMediaItem {
  mediaType?: GQLMediaType;
  reference?: string;
}

export const enum GQLMediaType {
  Image = 'Image',
  Video = 'Video'
}

export interface GQLPlanAttachment {
  planId: string;
}

export interface GQLMapResponse {
  friendPlans?: Array<GQLUserWithPlan | null>;
  publicPlans?: Array<GQLUserWithPlan | null>;
  colivings?: Array<GQLColivingResult | null>;
  coworkings?: Array<GQLCoworkingResult | null>;
}

export interface GQLUserWithPlan {
  user: GQLUser;
  plan: GQLPlan;
}

export interface GQLChannelsResult {
  channel?: GQLChannel;
  latestMessage?: GQLMessage;
  members?: Array<GQLUser | null>;
  unreadMessageCount?: number;
}

export interface GQLChannel {
  id?: string;
  name?: string;
  createdAt?: string;
  membersHash?: string;
  chatType?: GQLChatType;
  isMuted?: boolean;
}

export const enum GQLChatType {
  DIRECT = 'DIRECT',
  GROUP = 'GROUP'
}

export interface GQLGetChannelsResponse {
  results?: Array<GQLChannelsResult | null>;
  page?: GQLPage;
}
// ...
export interface QueryToChannelsArgs {
  page?: GQLPageInput;
}